Refactor SSOKeyCard func's into components
All the func's building the sub-components
were prime targets for component's with props.

CORL-929
