+++ src/core/client/admin/routes/Configure/sections/Auth/SSOKeyRotation/RotationDropdown.css
@@ -1,0 +1,3 @@
+.rotate {
+  margin-right: var(--v2-spacing-1)
+}
