+++ src/core/client/admin/routes/Configure/sections/Auth/SSOKeyRotation/SSOKeyCard.tsx
@@ -1,5 +1,4 @@
 import { Localized } from "@fluent/react/compat";
-import cn from "classnames";
 import React, { FunctionComponent, useCallback } from "react";
 import CopyToClipboard from "react-copy-to-clipboard";
 
@@ -7,39 +6,23 @@ import { useMutation } from "coral-framework/lib/relay";
 import {
   Button,
   Card,
-  ClickOutside,
-  Dropdown,
-  DropdownButton,
   Flex,
   HorizontalGutter,
   Icon,
   Label,
   PasswordField,
-  Popover,
   TextField,
-  Tooltip,
-  TooltipButton,
 } from "coral-ui/components/v2";
 
+import DateField from "./DateField";
 import DeactivateSSOKeyMutation from "./DeactivateSSOKeyMutation";
 import DeleteSSOKeyMutation from "./DeleteSSOKeyMutation";
 import RotateSSOKeyMutation from "./RotateSSOKeyMutation";
+import RotationDropDown from "./RotationDropdown";
+import { RotateOptions } from "./RotationOption";
+import StatusField, { SSOKeyStatus } from "./StatusField";
 
 import styles from "./SSOKeyCard.css";
-
-export enum SSOKeyStatus {
-  EXPIRED,
-  EXPIRING,
-  ACTIVE,
-}
-
-export enum RotateOptions {
-  NOW = "NOW",
-  IN10SECONDS = "IN10SECONDS",
-  IN1DAY = "IN1DAY",
-  IN1WEEK = "IN1WEEK",
-  IN30DAYS = "IN30DAYS",
-}
 
 export interface SSOKeyDates {
   readonly createdAt: string;
@@ -55,259 +38,14 @@ interface Props {
   dates: SSOKeyDates;
 }
 
-function getStatusField(status: SSOKeyStatus) {
-  if (status === SSOKeyStatus.ACTIVE) {
-    return (
-      <Localized id="configure-auth-sso-rotate-statusActive">
-        <span className={cn(styles.status, styles.active)}>Active</span>
-      </Localized>
-    );
-  }
-  if (status === SSOKeyStatus.EXPIRING) {
-    return (
-      <Flex alignItems="center" justifyContent="center">
-        <Flex
-          alignItems="center"
-          justifyContent="center"
-          className={cn(styles.status, styles.expiring)}
-        >
-          <Icon className={styles.icon}>alarm</Icon>
-          <Localized id="configure-auth-sso-rotate-statusExpiring">
-            <span>Expiring</span>
-          </Localized>
-        </Flex>
-        <Tooltip
-          id="configure-auth-sso-rotate-expiringTooltip"
-          title=""
-          body={
-            <Localized id="configure-auth-sso-rotate-expiringTooltip">
-              <span>
-                An SSO key is expiring when it is scheduled for rotation.
-              </span>
-            </Localized>
-          }
-          button={({ toggleVisibility, ref, visible }) => (
-            <Localized
-              id="configure-auth-sso-rotate-expiringTooltip-toggleButton"
-              attrs={{ "aria-label": true }}
-            >
-              <TooltipButton
-                active
-                aria-label="Toggle expiring tooltip visibility"
-                toggleVisibility={toggleVisibility}
-                ref={ref}
-              />
-            </Localized>
-          )}
-        />
-      </Flex>
-    );
-  }
-  if (status === SSOKeyStatus.EXPIRED) {
-    return (
-      <Flex alignItems="center" justifyContent="center">
-        <Localized id="configure-auth-sso-rotate-statusExpired">
-          <span className={cn(styles.status, styles.expired)}>Expired</span>
-        </Localized>
-        <Tooltip
-          id="configure-auth-sso-rotate-expiredTooltip"
-          title=""
-          body={
-            <Localized id="configure-auth-sso-rotate-expiredTooltip">
-              <span>
-                An SSO key is expired when it has been rotated out of use.
-              </span>
-            </Localized>
-          }
-          button={({ toggleVisibility, ref, visible }) => (
-            <Localized
-              id="configure-auth-sso-rotate-expiredTooltip-toggleButton"
-              attrs={{ "aria-label": true }}
-            >
-              <TooltipButton
-                active
-                aria-label="Toggle expired tooltip visibility"
-                toggleVisibility={toggleVisibility}
-                ref={ref}
-              />
-            </Localized>
-          )}
-        />
-      </Flex>
-    );
-  }
-
-  return (
-    <Localized id="configure-auth-sso-rotate-statusUnknown">
-      <span>Unknown</span>
-    </Localized>
-  );
-}
-
-function getDateField(status: SSOKeyStatus, dates: SSOKeyDates) {
-  // Active
-  if (status === SSOKeyStatus.ACTIVE) {
-    return (
-      <>
-        <div className={styles.label}>
-          <Localized id="configure-auth-sso-rotate-activeSince">
-            <Label>Active Since</Label>
-          </Localized>
-        </div>
-        <Localized
-          id="configure-auth-sso-rotate-date"
-          $date={new Date(dates.createdAt)}
-        >
-          <span className={styles.date}>{dates.createdAt}</span>
-        </Localized>
-      </>
-    );
-  }
-
-  // Expiring
-  if (status === SSOKeyStatus.EXPIRING) {
-    return (
-      <>
-        <div className={styles.label}>
-          <Localized id="configure-auth-sso-rotate-inactiveAt">
-            <Label>Inactive At</Label>
-          </Localized>
-        </div>
-        <Flex
-          alignItems="center"
-          justifyContent="center"
-          className={styles.date}
-        >
-          <Localized
-            id="configure-auth-sso-rotate-date"
-            $date={
-              dates.inactiveAt
-                ? new Date(dates.inactiveAt)
-                : new Date(dates.createdAt)
-            }
-          >
-            {dates.inactiveAt}
-          </Localized>
-        </Flex>
-      </>
-    );
-  }
-
-  // Expired
-  if (status === SSOKeyStatus.EXPIRED) {
-    return (
-      <>
-        <div className={styles.label}>
-          <Localized id="configure-auth-sso-rotate-inactiveSince">
-            <Label>Inactive Since</Label>
-          </Localized>
-        </div>
-        <Flex
-          alignItems="center"
-          justifyContent="center"
-          className={styles.date}
-        >
-          <Localized
-            id="configure-auth-sso-rotate-date"
-            $date={
-              dates.inactiveAt
-                ? new Date(dates.inactiveAt)
-                : new Date(dates.createdAt)
-            }
-          >
-            {dates.inactiveAt}
-          </Localized>
-        </Flex>
-      </>
-    );
-  }
-
-  return null;
-}
-
-function getTranslationForRotationOption(r: string) {
-  switch (r) {
-    case RotateOptions.NOW: {
-      return <Localized id="configure-auth-sso-rotate-now">Now</Localized>;
-    }
-    case RotateOptions.IN10SECONDS: {
-      return (
-        <Localized id="configure-auth-sso-rotate-10seconds">
-          10 seconds from now
-        </Localized>
-      );
-    }
-    case RotateOptions.IN1DAY: {
-      return (
-        <Localized id="configure-auth-sso-rotate-1day">
-          1 day from now
-        </Localized>
-      );
-    }
-    case RotateOptions.IN1WEEK: {
-      return (
-        <Localized id="configure-auth-sso-rotate-1week">
-          1 week from now
-        </Localized>
-      );
-    }
-    case RotateOptions.IN30DAYS: {
-      return (
-        <Localized id="configure-auth-sso-rotate-30days">
-          30 days from now
-        </Localized>
-      );
-    }
-    default:
-      return <Localized id="configure-auth-sso-rotate-now">Now</Localized>;
-  }
-}
-
-function getActionButton(
+function createActionButton(
   status: SSOKeyStatus,
   onRotateKey: (rotation: string) => void,
   onDeactivateKey: () => void,
   onDelete: () => void
 ) {
   if (status === SSOKeyStatus.ACTIVE) {
-    return (
-      <Localized
-        id="configure-auth-sso-rotate-dropdown-description"
-        attrs={{ description: true }}
-      >
-        <Popover
-          id="sso-key-rotate"
-          placement="bottom-start"
-          description="A dropdown to rotate the SSO key"
-          body={({ toggleVisibility }) => (
-            <ClickOutside onClickOutside={toggleVisibility}>
-              <Dropdown>
-                {Object.keys(RotateOptions).map((r: string) => (
-                  <DropdownButton
-                    key={r}
-                    onClick={() => {
-                      onRotateKey(r);
-                      toggleVisibility();
-                    }}
-                  >
-                    {getTranslationForRotationOption(r)}
-                  </DropdownButton>
-                ))}
-              </Dropdown>
-            </ClickOutside>
-          )}
-        >
-          {({ toggleVisibility, ref, visible }) => (
-            <Button onClick={toggleVisibility} ref={ref} color="regular">
-              <Localized id="configure-auth-sso-rotate-rotate">
-                <span className={styles.action}>Rotate</span>
-              </Localized>
-              <Icon>arrow_drop_down</Icon>
-            </Button>
-          )}
-        </Popover>
-      </Localized>
-    );
+    return <RotationDropDown onRotateKey={onRotateKey} />;
   }
   if (status === SSOKeyStatus.EXPIRING) {
     return (
@@ -428,11 +166,13 @@ const SSOKeyCard: FunctionComponent<Props> = ({
                   <Label>Status</Label>
                 </Localized>
               </div>
-              {getStatusField(status)}
+              <StatusField status={status}></StatusField>
             </div>
-            <div>{getDateField(status, dates)}</div>
+            <div>
+              <DateField status={status} dates={dates} />
+            </div>
           </Flex>
-          {getActionButton(status, onRotate, onDeactivate, onDelete)}
+          {createActionButton(status, onRotate, onDeactivate, onDelete)}
         </Flex>
       </HorizontalGutter>
     </Card>
